% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphitr.R
\name{format_host_name}
\alias{format_host_name}
\title{Formats a hostname so that it fullfills the following scheme:}
\usage{
format_host_name(host, default_scheme = "http")
}
\arguments{
\item{host}{The hostname (can also be a local path to a file or anything that read.csv supports)}

\item{default_scheme}{The scheme to append if no scheme is found}
}
\value{
The formatted host string
}
\description{
<scheme>://<host>{/}
}
\details{
where scheme should be one of 'http' or 'file'. If the scheme is
'file' then no trailing slash is appended
}
