% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphitr.R
\name{graphite_get}
\alias{graphite_get}
\title{Get raw metrics from Graphite.}
\usage{
graphite_get(host, from, until, target, format = "csv")
}
\arguments{
\item{host}{The address of the Graphite server}

\item{from}{Start of the time series. You can use either a Date or a string}

\item{until}{End of the time series. You can either use a Data or a string}

\item{target}{The metric to retrieve}

\item{format}{The format of the data. Defaults to csv}
}
\description{
Get raw metrics from Graphite.
}
