% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphitr.R
\name{graphite_render_uri}
\alias{graphite_render_uri}
\title{Render the graphite URL from multiple parameters.}
\usage{
graphite_render_uri(host, from, until, target, format = "csv", ...)
}
\arguments{
\item{host}{The address of the graphite server}

\item{from}{Start of the time series}

\item{until}{End of the time series}

\item{format}{The output format}
}
\value{
The rendered URI
}
\description{
Render the graphite URL from multiple parameters.
}
